edited destroy action in reviews_controller to conditionally display a message if review cannot be deleted

+++ wood-tavern-backend/app/controllers/reviews_controller.rb
@@ -41,7 +41,12 @@ class ReviewsController < ApplicationController
 
   # DELETE /reviews/1
   def destroy
-    review.destroy
+    review = Review.find_by(id: params[:id])
+    if review
+      review.destroy
+    else
+      render json: { message: 'Review not found' }
+    end
   end
 
   private
